+++ rex/rextransition.h
@@ -55,17 +55,17 @@ typedef struct rex_transition_s {
 	unsigned int type;
 	rexchar_t lowin;
 	rexchar_t highin;
-	long srcuid;
-	long dstuid;
+	size_t srcuid;
+	size_t dstuid;
 } rex_transition_t;
 
 
-rex_transition_t *rex_transitions_add(rarray_t *trans, rexchar_t c1, rexchar_t c2, unsigned long srcuid, unsigned long dstuid);
-rex_transition_t *rex_transitions_add_e(rarray_t *etrans, unsigned long srcuid, unsigned long dstuid);
+rex_transition_t *rex_transitions_add(rarray_t *trans, rexchar_t c1, rexchar_t c2, size_t srcuid, size_t dstuid);
+rex_transition_t *rex_transitions_add_e(rarray_t *etrans, size_t srcuid, size_t dstuid);
 void rex_transitions_normalize(rarray_t *trans);
-void rex_transitions_negative(rarray_t *dtrans, rarray_t *strans, unsigned long srcuid, unsigned long dstuid);
+void rex_transitions_negative(rarray_t *dtrans, rarray_t *strans, size_t srcuid, size_t dstuid);
 rex_transition_t *rex_transitions_find(rarray_t *trans, rexchar_t c);
-void rex_transitions_renamedestination(rarray_t *trans, long dstold, long dstnew);
+void rex_transitions_renamedestination(rarray_t *trans, size_t dstold, size_t dstnew);
 void rex_transitions_dump(rarray_t *trans);
 
 #ifdef __cplusplus
